backend/requirements.txt
asgiref==3.2.10
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
django-colorfield==0.7.0
djoser==2.1.0
drf-base64==2.0
drf-yasg==1.20.0
python-dotenv==0.19.2
psycopg2-binary==2.8.6
pytz==2020.1
Pillow==9.2.0
sqlparse==0.3.1
gunicorn==20.0.4